refactor: move EmptyScreen/FilteredEmptyScreen to ui package
Fixes #6919
