+++ packages/renderer/src/lib/extensions/ExtensionDetailsReadme.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 import { faFileText } from '@fortawesome/free-solid-svg-icons';
+import { EmptyScreen } from '@podman-desktop/ui-svelte';
 import { onMount } from 'svelte';
 
 import Markdown from '../markdown/Markdown.svelte';
-import EmptyScreen from '../ui/EmptyScreen.svelte';
 
 export let readme: { content?: string; uri?: string };
