+++ packages/renderer/src/lib/service/ServiceEmptyScreen.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
+import { EmptyScreen } from '@podman-desktop/ui-svelte';
+
 import ServiceIcon from '../images/ServiceIcon.svelte';
-import EmptyScreen from '../ui/EmptyScreen.svelte';
 </script>
 
 <EmptyScreen icon="{ServiceIcon}" title="No services" message="Try switching to a different context or namespace" />
