+++ packages/renderer/src/lib/volume/VolumeEmptyScreen.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
+import { EmptyScreen } from '@podman-desktop/ui-svelte';
+
 import VolumeIcon from '../images/VolumeIcon.svelte';
-import EmptyScreen from '../ui/EmptyScreen.svelte';
 </script>
 
 <EmptyScreen
