+++ packages/ui/src/lib/index.ts
@@ -26,6 +26,8 @@ import Link from './link/Link.svelte';
 import Modal from './modal/Modal.svelte';
 import LinearProgress from './progress/LinearProgress.svelte';
 import Spinner from './progress/Spinner.svelte';
+import EmptyScreen from './screen/EmptyScreen.svelte';
+import FilteredEmptyScreen from './screen/FilteredEmptyScreen.svelte';
 import Tab from './tab/Tab.svelte';
 import TableDurationColumn from './table/DurationColumn.svelte';
 import TableSimpleColumn from './table/SimpleColumn.svelte';
@@ -38,7 +40,9 @@ export type { ButtonType };
 export {
   Button,
   Checkbox,
+  EmptyScreen,
   ErrorMessage,
+  FilteredEmptyScreen,
   Input,
   LinearProgress,
   Link,
