+++ packages/ui/src/lib/screen/EmptyScreen.svelte
@@ -1,9 +1,11 @@
 <script lang="ts">
 import { faPaste } from '@fortawesome/free-solid-svg-icons';
-import { isFontAwesomeIcon } from '@podman-desktop/ui-svelte';
 import { onMount } from 'svelte';
 import Fa from 'svelte-fa';
 
+import { isFontAwesomeIcon } from '../utils/icon-utils';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export let icon: any;
 export let title = 'No title';
 export let message = 'Message';
@@ -21,7 +23,7 @@ onMount(() => {
   processed = true;
 });
 
-function copyRunInstructionToClipboard() {
+function copyRunInstructionToClipboard(): void {
   const text = copyTextDivElement?.textContent;
   if (text) {
     window.clipboardWriteText(text);
@@ -60,7 +62,7 @@ let copyTextDivElement: HTMLDivElement;
           data-testid="copyTextDivElement">
           {commandline}
         </div>
-        <button title="Copy To Clipboard" class="ml-5" on:click="{() => copyRunInstructionToClipboard()}"
+        <button title="Copy To Clipboard" class="ml-5" on:click="{copyRunInstructionToClipboard}"
           ><Fa class="h-5 w-5 cursor-pointer text-xl text-purple-500 hover:text-purple-600" icon="{faPaste}" /></button>
       </div>
     {/if}
