+++ packages/ui/src/lib/screen/FilteredEmptyScreen.svelte
@@ -1,15 +1,16 @@
 <script lang="ts">
-import { Button } from '@podman-desktop/ui-svelte';
 import { createEventDispatcher } from 'svelte';
 
-import EmptyScreen from '../ui/EmptyScreen.svelte';
+import Button from '../button/Button.svelte';
+import EmptyScreen from './EmptyScreen.svelte';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export let icon: any;
 export let kind: string;
 export let searchTerm: string;
 
 const dispatch = createEventDispatcher();
-function onResetFilter() {
+function onResetFilter(): void {
   if (dispatch('resetFilter', searchTerm, { cancelable: true })) {
     searchTerm = '';
   }
@@ -23,5 +24,5 @@ $: filter = searchTerm && searchTerm.length > 20 ? 'filter' : `'${searchTerm}'`;
   title="No {kind} matching {filter} found"
   message="Not what you expected? Double-check your spelling."
   detail="Just want to view all of your {kind}?">
-  <Button on:click="{() => onResetFilter()}">Clear filter</Button>
+  <Button on:click="{onResetFilter}">Clear filter</Button>
 </EmptyScreen>
